# app/Http/Controllers/RuangController.php
<?php

namespace App\Http\Controllers;

use Illuminate\Http\Request;
use App\kelas;

class ruangController extends Controller
{
    /**
     * Display a listing of the resource.
     *
     * @return \Illuminate\Http\Response
     *
     * @SWG\Get(
     *      path="/api/v1/kelas",
     *      summary="Retrieves the collection of Ruang resources.",
     *      produces={"application/json"},
     *      tags={"Ruang"},
     *      @SWG\Response(
     *          response=200,
     *          description="Ruang collection.",
     *          @SWG\Schema(
     *               type="array",
     *               @SWG\Items(ref="#/definitions/kelas")
     *          )
     *      ),
     *      @SWG\Response(
     *          response=401,
     *          description="Unauthorized action.",
     *      ),
     *      @SWG\Parameter(
     *          name="Authorization",
     *          in="header",
     *          required=true,
     *          type="string"
     *      )
     * )
     */ 
    public function index()
    {
       
        $user=Auth::user();

        if($user->id_level!=3){

            return response()->json(['error'=>Auth::user()->name.',Forbidden'], 403);

        }else($user->id_level ==1) {

             $kelas=kelas::paginate(5);
            return response()->json($kelas->toArray());


    }

    /**
     * Show the form for creating a new resource.
     *
     * @return \Illuminate\Http\Response
     */
    public function create()
    {
        //
    }

    /**
    * @SWG\Post(
    *      path="/api/kelas",
    *      summary="Data Ruang",
    *      produces={"application/json"},
    *      consumes={"application/json"},
    *      tags={"Ruang"},
    *      @SWG\Response(
    *          response=200,
    *          description="Users Token.",
    *          @SWG\Property(
    *              property="token",
    *              type="string"
    *          )
    *      ),
    *      @SWG\Response(
    *          response=401,
    *          description="Unauthorized action.",
    *      ),
    *      @SWG\Parameter(
    *          name="Authorization",
    *          in="header",
    *          required=true,
    *          type="string"
    *      ),
    *      @SWG\parameter(
    *          name="Data Kelas",
    *          in="body",
    *          required=true,
    *          type="string",
    *          @SWG\Schema(
    *              type="string"
    *          )
    *      )
    * )
   */
    public function store(Request $request)
    {
        $this->validate($request,[
            'nama'=>'required',
            ]);
        $kelas = new kelas();
        $kelas->nama=$request->input('nama');
        $kelas->save();
        return response()->json($kelas);
    }

    /**
     * Display a listing of the resource.
     *
     * @return \Illuminate\Http\Response
     *
     * @SWG\Get(
     *      path="/api/v1/kelas/{id}",
     *      summary="Retrieves the collection of Ruang resources.",
     *      produces={"application/json"},
     *      tags={"Ruang"},
     *      @SWG\Response(
     *          response=200,
     *          description="ruang collection.",
     *          @SWG\Schema(
     *               type="array",
     *               @SWG\Items(ref="#/definitions/kelas")
     *          )
     *      ),
     *      @SWG\Response(
     *          response=401,
     *          description="Unauthorized action.",
     *      ),
     *      @SWG\Parameter(
     *          name="Authorization",
     *          in="header",
     *          required=true,
     *          type="string"
     *      ),
     *      @SWG\parameter(
     *          name="id",
     *          in="path",
     *          required=true,
     *          type="integer"
     *      )
     *  )
     */
    public function show($id)
    {
        $kelas = kelas::find($id);
        return response()->json($kelas->toArray());
    }

    /**
     * Show the form for editing the specified resource.
     *
     * @param  int  $id
     * @return \Illuminate\Http\Response
     */
    public function edit($id)
    {
        //
    }

    /**
     * @SWG\Put(
    *      path="/api/v1/kelas/{id}",
    *      summary="Data Ruang",
    *      produces={"application/json"},
    *      consumes={"application/json"},
    *      tags={"Ruang"},
    *      @SWG\Response(
    *          response=200,
    *          description="Success Update.",
    *          @SWG\Property(
    *              property="token",
    *              type="string"
    *          )
    *      ),
    *      @SWG\Response(
    *          response=401,
    *          description="Unauthorized action.",
    *      ),
    *      @SWG\Parameter(
    *          name="Authorization",
    *          in="header",
    *          required=true,
    *          type="string"
    *      ),
    *       @SWG\parameter(
    *          name="id",
    *          in="path",
    *          required=true,
    *          type="integer"
    *      ),
    *      @SWG\parameter(
    *          name="Data Ruang",
    *          in="body",
    *          required=true,
    *          type="string",
    *          @SWG\Schema(
    *              type="string"
    *          )
    *      )
    * )
   */
    public function update(Request $request, $id)
    {
        $kelas = kelas::find($id);
        $kelas->nama=$request->input('nama');
        $kelas->save();
        return response()->json($kelas->toArray());
    }

    /**
     * Remove the specified resource from storage.
     *
     * @param  int  $id
     * @return \Illuminate\Http\Response
     *
     * @SWG\Delete(
     *      path="/api/v1/kelas/{id}",
     *      summary="Removes the Ruang resource.",
     *      produces={"application/json"},
     *      tags={"Ruang"},
     *      @SWG\Response(
     *          response=204,
     *          description="Ruang resource delete."
     *      ),
     *      @SWG\Response(
     *          response=401,
     *          description="Unauthorized Action."
     *      ),
     *      @SWG\Response(
     *          response=404,
     *          description="Resource not found."
     *      ),
     *      @SWG\Parameter(
     *          name="Authorization",
     *          in="header",
     *          required=true,
     *          type="string"
     *      ),
     *      @SWG\parameter(
     *          name="id",
     *          in="path",
     *          required=true,
     *          type="integer"
     *      )
     *  )
     */
    public function destroy($id)
    {
        $kelas = kelas::find($id);
        $kelas->delete();
        return response()->json($kelas->toArray());
    }
}