# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Sample API",
        "contact": {
            "name": "Erik Alexander, Friscilia Sutanto, Sri Wardani",
            "url": "https://uph.edu"
        },
        "version": "1.0"
    },
    "host": "webservice.app",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/v1/kelas": {
            "get": {
                "tags": [
                    "Ruang"
                ],
                "summary": "Retrieves the collection of Ruang resources.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ruang collection.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/kelas"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/api/kelas": {
            "post": {
                "tags": [
                    "Ruang"
                ],
                "summary": "Data Ruang",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Data Kelas",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users Token.",
                        "schema": {
                            "property": "token",
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/api/v1/kelas/{id}": {
            "get": {
                "tags": [
                    "Ruang"
                ],
                "summary": "Retrieves the collection of Ruang resources.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ruang collection.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/kelas"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "put": {
                "tags": [
                    "Ruang"
                ],
                "summary": "Data Ruang",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Data Ruang",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Update.",
                        "schema": {
                            "property": "token",
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ruang"
                ],
                "summary": "Removes the Ruang resource.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Ruang resource delete."
                    },
                    "401": {
                        "description": "Unauthorized Action."
                    },
                    "404": {
                        "description": "Resource not found."
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "kelas": {
            "properties": {
                "nama": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                }
            }
        },
...
    }
}